Function cjdns_core::splice::routes_through

pub fn routes_through<L: LabelBits>(
    destination: RoutingLabel<L>, 
    mid_path: RoutingLabel<L>
) -> bool

This will return true if the node at the end of the route given by mid_path is a hop along the path given by destination.

assert_eq!(routes_through(l("0000.001b.0535.10e5"), l("0000.0000.0000.0015")), true);
assert_eq!(routes_through(l("0000.001b.0535.10e5"), l("0000.0000.0000.0013")), false);

See: LabelSplicer_routesThrough()